Update Fourier_approximator.py
Fixed the Fourier coefficients!
My problem was less about the code and more about the math. A large part of the Fourier Series is a change of variable in the integral-- something which, when forgotten about, did not matter for simple demonstrations but for an actually useful program it caused a lot of headache. Lesson learned!